epiphany: Implement reset ioctl
This is potentially unsafe since another thread might be accessing the
emesh concurrently. However, we need to provide this functionality to
user-space and this is the proper API.

So make it a requirement that the calling program must not access the
emesh while the reset is underway for now. Revisit and fix later.

Signed-off-by: Ola Jeppsson <ola@adapteva.com>