epiphany: Allow FPGA elink regs to be inside mesh region
The FPGA elink registers have moved inside the mesh region in OH, so we
need to support that in the driver.

We might want replace the elink 'reg' property in the device-tree with a
'coreid' property for the elink instead, as it might feel more natural.
But this will work for now.

Signed-off-by: Ola Jeppsson <ola@adapteva.com>