epiphany: Move power supply from array to elink
Background: The FPGA elink design uses a clock from the chip it's
connected to. That clock is needed even for accessing the FPGA
registers only (which we need for probing). I can't find a good way to
express that connection in the device tree (there would be a cyclic
dependency), and then we'd also have to implement a clock source driver
for the chip clock, which I really think is unnecessary now.

Moving the supply from the array to the elink is the easiest way to
solve this.

Signed-off-by: Ola Jeppsson <ola@adapteva.com>