epiphany: Try to work around some mailbox issues
This change (almost?) mitigates some weirdness with the FPGA elink.
Sometimes there's an Imprecise external abort when reading from FPGA
elink registers (triggered by the timeout circuit in the elink).

The chain of events is something resembling this:

Clock Event
0.    Mailbox is empty, mailbox interrupt is disabled
1.    Read  to ELINK_MAILBOXSTAT from host
2.    Write to ELINK_MAILBOXLO from Epiphany
3.    Read  to ELINK_RXCFG from host
[ external abort here ??? ]
4.    Write to ELINK_RXCFG from host that enables mailbox irq
5.    Read  to ELINK_MAILBOXSTAT from host interrupt handler
6.    Read  to ELINK_MAILBOXSTAT from host
7.    Read  to ELINK_RXCFG from host irq handler
8.    Write to ELINK_RXCFG from host irq handler that disables mailbox irq.
9.    Read  to ELINK_MAILBOXSTAT from host

In link_char_ioctl_mailbox_read(), wait_event_interruptible() used
elink_mailbox_empty_p() as condition, without holding the
&epiphany.driver_lock mutex. The locking scheme requires that the driver
lock must be held when accessing FPGA elink registers.

Add an atomic conditional, mailbox_maybe_not_empty, to the elink_device
structure. It is then toggled in the irq handler, so it can be used as a
conditional in the ioctl call.

Also, add a timeout to the wait event call. This shouldn't be necessary
but the driver seems to occasionally get stuck without it. Could also be
H/W related.

Signed-off-by: Ola Jeppsson <ola@adapteva.com>