epiphany: Translate pfn in page fault handler
Move translation of mesh pfn from _elink_char_mmap() to
epiphany_vm_fault().

Implement mesh_pfn_to_phys_pfn() that translates a mesh pfn to its
physical page frame, or returns an error. Use this in the fault handler
to verify whether a page should be allowed to be mapped in.

Signed-off-by: Ola Jeppsson <ola@adapteva.com>