serial: xuartps: Add quirk for floating RX pin
The Parallella board does not have a pull-up resistor connected to its
serial RX pin, so the pin is in a floating state when no serial cable is
attached. When boot console is enabled, this commonly leads to the
driver getting stuck in an endless loop reading the RX FIFO which never
empties.

This patch adds a boolean device-tree property for the xuartps driver:
"cdns,rx-no-pullup"

With the above property set the driver adds a limit to at most 1.5x RX
FIFO depth reads from the RX FIFO in the same loop. If it goes above
that 'unconnected floating RX pin quirk' is detected and the driver
disables receive on the port permanently until reboot / module unload.

Tested 200x reboot cycles on 2 different boards.

Signed-off-by: Ola Jeppsson <ola@adapteva.com>