epiphany: Use correct supply name in call to regulator_get_optional
In elink_of_probe_supplies():
Remove the logic that extracts the regulator name from the regulator
node. This is not how it should be done.
We should just pass in "vdd" to regulator_get(), and the regulator core
will look up "vdd-supply" in our dt node and figure out the rest.

We were relying on the fallback method previously, which doesn't check
for dependencies, so if the regulator isn't ready it will still return
-ENODEV.

Now regulator_get() will return -EPROBE_DEFER instead of -ENODEV if the
regulator isn't ready yet.

Signed-off-by: Ola Jeppsson <ola@adapteva.com>