epiphany: Fix deadlocks
We cannot call synchronize_rcu_expedited() from epiphany_vm_open() while
holding a mmap_sem lock because it can deadlock with
epiphany_vm_freeze() if another thread is traversing the vma list, still
holding the rcu lock.

Since we're holding a writer lock on mmap_sem in epiphany_vm_open() this
is not a good place to release the lock. Instead defer adding the new
vma to the vma_list to the fault handler.

Signed-off-by: Ola Jeppsson <ola@adapteva.com>