epiphany: Add back epiphany_vm_open
We need to have a vma open operation so we can track all vma's that
references epiphany memory.

brk() / mremap() / munmap() + probable other syscalls can result in new
vma's so it's not sufficient to only add new vma's in mmap().

Other drivers seem to set vma->vm_private_data in mmap() instead of
vm_open() but that doesn't make sense for us as long as we only need to
be able to access the epiphany_vma_entry struct from vm operations. We
can just use the private data to point to the epiphany_vma_entry so we
don't have to walk the vma_list every time we need it.

Signed-off-by: Ola Jeppsson <ola@adapteva.com>